Introduce OpenAPI based Org VDC network management functions for NSX-V and NSX-T (#354)
This PR introduces:
* OpenApiOrgVdcNetwork structure and type meant to handle Org VDC networks for both NSX-T and NSX-V backed.
* Structure NsxtImportableSwitch meant for lookup "importable switches" (NSX-T segments) to be used for NSX-T ( Imported networks. Methods GetNsxtImportableSwitchByName and GetAllNsxtImportableSwitches aid that
* Methods vdc.IsNsxt and vdc.IsNsxv help to verify which type of network provider this VDC is backed.

Additionally it ads:

* nsxtVdc shorthand in TestVCD struct for testing
* Two new config fields edgeGateway (to use existing edge gateway) and unusedSegment which is needed to test our Imported NSX-T Org VDC network
* Converts VirtualQuantity int->int64 to avoid variable overflow with big sizes (vmware/terraform-provider-vcd#626)

+++ govcd/vapp_test.go
@@ -143,7 +143,7 @@ func (vcd *TestVCD) Test_ChangeCPUCountWithCore(check *C) {
 		check.Skip("Skipping test because vApp was not successfully created at setup")
 	}
 
-	currentCpus := 0
+	currentCpus := int64(0)
 	currentCores := 0
 
 	// save current values
@@ -158,8 +158,8 @@ func (vcd *TestVCD) Test_ChangeCPUCountWithCore(check *C) {
 	}
 
 	cores := 2
-	cpuCount := 4
-	task, err := vcd.vapp.ChangeCPUCountWithCore(cpuCount, &cores)
+	cpuCount := int64(4)
+	task, err := vcd.vapp.ChangeCPUCountWithCore(int(cpuCount), &cores)
 	check.Assert(err, IsNil)
 	err = task.WaitTaskCompletion()
 	check.Assert(err, IsNil)
@@ -181,7 +181,7 @@ func (vcd *TestVCD) Test_ChangeCPUCountWithCore(check *C) {
 	}
 
 	// return tu previous value
-	task, err = vcd.vapp.ChangeCPUCountWithCore(currentCpus, &currentCores)
+	task, err = vcd.vapp.ChangeCPUCountWithCore(int(currentCpus), &currentCores)
 	check.Assert(err, IsNil)
 	err = task.WaitTaskCompletion()
 	check.Assert(err, IsNil)
